Reduce similarity threshold for beets to autotag
I find that when you download songs on YouTube they typically have (Audio) or (Official Video) in the title, and that brings the similarity to the actual album down, below what beets will automatically accept. So, I've reduced beets shyness, and instructed it to accept changes when the downloaded album is over 90% similar to the real one.